+++ clang-tools-extra/test/clang-tidy/checkers/daedalean-local-methods-and-types.hh
@@ -1,0 +1,8 @@
+// RUN: %check_clang_tidy %s daedalean-local-methods-and-types %t
+
+void foo1();
+void foo2();
+
+struct S {
+
+};